activemq/company/requirements.txt
stomp.py==6.0
CherryPy
babel
humanize
